# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PikPakGo API",
        "description": "PikPakGo Travel Marketplace API - Hotels, Vacation Rentals & More",
        "contact": {
            "name": "PikPakGo Support",
            "email": "reservations@pikpakgo.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://pickpackgo.in-sourceit.com/api",
            "description": "Production Server"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/payments/guest/process": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment for a guest booking",
                "operationId": "4e0c02a0a13ffde4d5073403bdadfdeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_reference",
                                    "payment_method",
                                    "card_number",
                                    "card_holder_name",
                                    "card_expiry_month",
                                    "card_expiry_year",
                                    "card_cvv",
                                    "billing_first_name",
                                    "billing_last_name",
                                    "billing_email",
                                    "billing_address",
                                    "billing_city",
                                    "billing_country",
                                    "billing_postal_code"
                                ],
                                "properties": {
                                    "booking_reference": {
                                        "type": "string",
                                        "example": "PKG-ABCD1234"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "debit_card"
                                        ]
                                    },
                                    "card_number": {
                                        "type": "string",
                                        "example": "4111111111111111"
                                    },
                                    "card_holder_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "card_expiry_month": {
                                        "type": "string",
                                        "example": "12"
                                    },
                                    "card_expiry_year": {
                                        "type": "string",
                                        "example": "2027"
                                    },
                                    "card_cvv": {
                                        "type": "string",
                                        "example": "123"
                                    },
                                    "billing_first_name": {
                                        "type": "string"
                                    },
                                    "billing_last_name": {
                                        "type": "string"
                                    },
                                    "billing_email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "billing_address": {
                                        "type": "string"
                                    },
                                    "billing_city": {
                                        "type": "string"
                                    },
                                    "billing_country": {
                                        "type": "string"
                                    },
                                    "billing_postal_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful"
                    },
                    "400": {
                        "description": "Payment failed or already paid"
                    }
                }
            }
        },
        "/payments/process": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Process payment for an authenticated user booking",
                "operationId": "332caf4b9a74a62bf52c2a4b2aa19659",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_reference",
                                    "payment_method",
                                    "card_number",
                                    "card_holder_name",
                                    "card_expiry_month",
                                    "card_expiry_year",
                                    "card_cvv"
                                ],
                                "properties": {
                                    "booking_reference": {
                                        "type": "string",
                                        "example": "PKG-ABCD1234"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "credit_card",
                                            "debit_card"
                                        ]
                                    },
                                    "card_number": {
                                        "type": "string",
                                        "example": "4111111111111111"
                                    },
                                    "card_holder_name": {
                                        "type": "string"
                                    },
                                    "card_expiry_month": {
                                        "type": "string",
                                        "example": "12"
                                    },
                                    "card_expiry_year": {
                                        "type": "string",
                                        "example": "2027"
                                    },
                                    "card_cvv": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment successful"
                    },
                    "400": {
                        "description": "Payment failed"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/{transactionId}/status": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get payment transaction status",
                "operationId": "1b76b1e740091eecb60ebd27a9695db0",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction status"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/payments/guest/{transactionId}/status": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get guest payment transaction status",
                "operationId": "d800fadd9dc3f3a77bdb9fe72a0b979c",
                "parameters": [
                    {
                        "name": "transactionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction status"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/payments/history": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Get authenticated user's payment history",
                "operationId": "0f3bba38e2567507c8841330f3a539ac",
                "responses": {
                    "200": {
                        "description": "Paginated payment history"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}